#84 - add active scope to keyword model - add check keyword rankings commands - add shorcuts to actions inside of keyword and ranking models - update CheckRankingCommand.php to use action - add an exception to GetOrganicResults.php so we know when it doesn't return any data - add setActive method to Keyword, which updates appropriate date fields
