+++ src/Models/Keyword.php
@@ -6,6 +6,7 @@ namespace Tipoff\Seo\Models;
 
 use Illuminate\Database\Eloquent\Builder;
 use Tipoff\Seo\Enum\KeywordType;
+use Tipoff\Seo\Services\Keyword\CheckRanking;
 use Tipoff\Support\Models\BaseModel;
 use Tipoff\Support\Traits\HasCreator;
 use Tipoff\Support\Traits\HasPackageFactory;
@@ -24,11 +25,11 @@ class Keyword extends BaseModel
         static::saving(function ($keyword) {
             $keyword->phrase = strtolower($keyword->phrase);
         });
+    }
 
-        static::addGlobalScope('active', function (Builder $builder) {
-            $builder->whereDate('keywords.tracking_requested_at', '<=', date('Y-m-d')) &&
-            ($builder->whereDate('keywords.tracking_stopped_at', '>=', date('Y-m-d')) || $builder->whereNull('keywords.tracking_stopped_at'));
-        });
+    public function scopeActive($query)
+    {
+        return $query->where('is_active', '=', true);
     }
 
     public function isBranded(): bool
@@ -45,12 +46,32 @@ class Keyword extends BaseModel
     {
         return $this->type == KeywordType::LOCAL;
     }
-  
+
+    public function setActive($status = true) : self
+    {
+        if ($status === true) {
+            $this->tracking_requested_at = now();
+        } else {
+            $this->tracking_stopped_at = now();
+        }
+
+        $this->is_active = $status;
+
+        $this->save();
+
+        return $this;
+    }
+
+    public function getRanking() : void
+    {
+        app(CheckRanking::class)($this);
+    }
+
     public function searchLocales()
     {
         return $this->belongsToMany(app('search_locales'))->withTimestamps();
     }
-      
+
     public function parent()
     {
         return $this->belongsTo(app('keyword'), 'parent_id');
