+++ tests/Unit/Services/Keyword/CheckRankingTest.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tipoff\Seo\Tests\Unit\Services\Ranking;
+
+use Tipoff\Seo\Tests\TestCase;
+
+class CheckRankingTest extends TestCase
+{
+    public function check_local_results()
+    {
+        return $this->assertTrue(true);
+    }
+
+    public function check_global_results()
+    {
+        return $this->assertTrue(true);
+    }
+}
